requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0